--- app/views/samples/index.html.erb
@@ -1,2 +1,0 @@
-<h1>Samples#index</h1>
-<p>Find me in app/views/samples/index.html.erb</p>
