--- app/views/samples/show.html.erb
@@ -1,1 +1,0 @@
-show.html.erb
