+++ db/migrate/20181204145612_create_receptions.rb
@@ -1,0 +1,8 @@
+class CreateReceptions < ActiveRecord::Migration[5.2]
+  def change
+    create_table :receptions do |t|
+
+      t.timestamps
+    end
+  end
+end
