+++ test/models/reception_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class ReceptionTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
